build: update go-client with paygate KYC fields

+++ cmd/apitest/transfer.go
@@ -26,7 +26,7 @@ func createDepository(ctx context.Context, api *moov.APIClient, u *user, account
 		HolderType:    "Individual",
 		Type:          account.Type,
 	}
-	dep, resp, err := api.DepositoriesApi.AddDepository(ctx, req, &moov.AddDepositoryOpts{
+	dep, resp, err := api.DepositoriesApi.AddDepository(ctx, u.ID, req, &moov.AddDepositoryOpts{
 		XIdempotencyKey: optional.NewString(generateID()),
 		XRequestID:      optional.NewString(requestID),
 	})
@@ -47,7 +47,7 @@ func createDepository(ctx context.Context, api *moov.APIClient, u *user, account
 
 func verifyDepository(ctx context.Context, api *moov.APIClient, accountID string, dep moov.Depository, u *user, requestID string) error {
 	// start micro deposits
-	resp, err := api.DepositoriesApi.InitiateMicroDeposits(ctx, dep.ID, &moov.InitiateMicroDepositsOpts{
+	resp, err := api.DepositoriesApi.InitiateMicroDeposits(ctx, u.ID, dep.ID, &moov.InitiateMicroDepositsOpts{
 		XIdempotencyKey: optional.NewString(generateID()),
 		XRequestID:      optional.NewString(requestID),
 	})
@@ -79,7 +79,7 @@ func verifyDepository(ctx context.Context, api *moov.APIClient, accountID string
 	}
 
 	// confirm micro deposits
-	resp, err = api.DepositoriesApi.ConfirmMicroDeposits(ctx, dep.ID, microDeposits, &moov.ConfirmMicroDepositsOpts{
+	resp, err = api.DepositoriesApi.ConfirmMicroDeposits(ctx, u.ID, dep.ID, microDeposits, &moov.ConfirmMicroDepositsOpts{
 		XIdempotencyKey: optional.NewString(generateID()),
 		XRequestID:      optional.NewString(requestID),
 	})
@@ -92,11 +92,10 @@ func verifyDepository(ctx context.Context, api *moov.APIClient, accountID string
 	return nil
 }
 
-func createOriginator(ctx context.Context, api *moov.APIClient, depId, requestID string) (moov.Originator, error) {
-	// Just get a valid date that's not today
+func createOriginator(ctx context.Context, api *moov.APIClient, depId, requestID, userID string) (moov.Originator, error) {
+	first, _ := name()
 	birthDate := time.Now().Truncate(24 * time.Hour).Add(-30 * 24 * time.Hour) // 30 days ago
 
-	first, _ := name()
 	req := moov.CreateOriginator{
 		DefaultDepository: depId,
 		Identification:    "123456789",
@@ -109,7 +108,7 @@ func createOriginator(ctx context.Context, api *moov.APIClient, depId, requestID
 		},
 		Metadata: fmt.Sprintf("%s Corp", first),
 	}
-	orig, resp, err := api.OriginatorsApi.AddOriginator(ctx, req, &moov.AddOriginatorOpts{
+	orig, resp, err := api.OriginatorsApi.AddOriginator(ctx, userID, req, &moov.AddOriginatorOpts{
 		XIdempotencyKey: optional.NewString(generateID()),
 		XRequestID:      optional.NewString(requestID),
 	})
@@ -138,7 +137,7 @@ func createReceiver(ctx context.Context, api *moov.APIClient, u *user, depId, re
 		},
 		Metadata: u.Name,
 	}
-	receiver, resp, err := api.ReceiversApi.AddReceivers(ctx, req, &moov.AddReceiversOpts{
+	receiver, resp, err := api.ReceiversApi.AddReceivers(ctx, u.ID, req, &moov.AddReceiversOpts{
 		XIdempotencyKey: optional.NewString(generateID()),
 		XRequestID:      optional.NewString(requestID),
 	})
@@ -151,7 +150,7 @@ func createReceiver(ctx context.Context, api *moov.APIClient, u *user, depId, re
 	return receiver, nil
 }
 
-func createTransfer(ctx context.Context, api *moov.APIClient, receiver moov.Receiver, orig moov.Originator, amount string, requestID string) (moov.Transfer, error) {
+func createTransfer(ctx context.Context, api *moov.APIClient, receiver moov.Receiver, orig moov.Originator, amount string, requestID, userID string) (moov.Transfer, error) {
 	req := moov.CreateTransfer{
 		TransferType:         "Push",
 		Amount:               amount,
@@ -172,7 +171,7 @@ func createTransfer(ctx context.Context, api *moov.APIClient, receiver moov.Rece
 		req.WEBDetail = createWEBDetail()
 
 	}
-	tx, resp, err := api.TransfersApi.AddTransfer(ctx, req, &moov.AddTransferOpts{
+	tx, resp, err := api.TransfersApi.AddTransfer(ctx, userID, req, &moov.AddTransferOpts{
 		XIdempotencyKey: optional.NewString(generateID()),
 		XRequestID:      optional.NewString(requestID),
 	})
@@ -184,7 +183,7 @@ func createTransfer(ctx context.Context, api *moov.APIClient, receiver moov.Rece
 	}
 	if *flagCleanup {
 		// Delete the transfer (and underlying file) since we're only making one Transfer
-		resp, err = api.TransfersApi.DeleteTransferByID(ctx, tx.ID, &moov.DeleteTransferByIDOpts{
+		resp, err = api.TransfersApi.DeleteTransferByID(ctx, userID, tx.ID, &moov.DeleteTransferByIDOpts{
 			XRequestID: optional.NewString(requestID),
 		})
 		if resp != nil {
